BJS2-33370 Правки после ревью
