+++ src/main/java/school/faang/abstraction/Archer.java
@@ -1,0 +1,13 @@
+package school.faang.abstraction;
+
+public class Archer extends Character {
+    public Archer(String name) {
+        super(name, 3, 10, 5);
+    }
+
+    @Override
+    public Character attack(Character character) {
+        character.health -= this.agility;
+        return character;
+    }
+}
