+++ src/main/java/school/faang/abstraction/Character.java
@@ -1,0 +1,21 @@
+package school.faang.abstraction;
+
+public abstract class Character {
+    protected String name;
+    protected int strength;
+    protected int agility;
+    protected int intelligence;
+    protected int health = 100;
+
+    public Character(String name) {
+        this.name = name;
+    }
+    public Character(String name, int strength, int agility, int intelligence) {
+        this.name = name;
+        this.strength = strength;
+        this.agility = agility;
+        this.intelligence = intelligence;
+    }
+
+    public abstract Character attack(Character character);
+}
