+++ src/main/java/school/faang/abstraction/Main.java
@@ -1,10 +1,10 @@
-package school.faang.BJS2_28997;
+package school.faang.abstraction;
 
 
 public class Main {
     public static void main(String[] args) {
-        Warrior warrior = new Warrior("Чарна");
-        Archer archer = new Archer("Валеска");
+        Character warrior = new Warrior("Чарна");
+        Character archer = new Archer("Валеска");
 
         warrior.attack(archer);
         System.out.printf("%s: %s%n", archer.name, archer.health);
