+++ src/main/java/school/faang/abstraction/Warrior.java
@@ -1,0 +1,13 @@
+package school.faang.abstraction;
+
+public class Warrior extends Character {
+    public Warrior(String name) {
+        super(name, 10, 5, 3);
+    }
+
+    @Override
+    public Character attack(Character character) {
+        character.health -= this.strength;
+        return character;
+    }
+}
